+++ apps/AssetBuilder/include/path.h
